+++ backend/controllers/user.js
@@ -16,21 +16,23 @@ exports.signup = (req, res, next) => {
 
     return res.status(400).json({ message: "Informations d'inscriptions manquantes." })
   }
+  console.log(req.body)
   db.user.findOne({ where: { email: req.body.email } })
     .then(value => {
       if (value === null) {
-        console.log("kutfkutf");
         if (!new RegExp(/^[a-zA-Z0-9.-_]+[@]{1}(groupomania)[.]{1}(fr)$/).test(req.body.email) ||
           !new RegExp(/(?!^[0-9]*$)(?!^[a-zA-Z]*$)^([a-zA-Z0-9]{8,32})$/).test(req.body.pwd)) {
-          res.status(400).json({ message: "Erreur d'adresse mail ou de mot de passe." })
+          console.log(!new RegExp(/^[a-zA-Z0-9.-_]+[@]{1}(groupomania)[.]{1}(fr)$/).test(req.body.email))
+          console.log(!new RegExp(/(?!^[0-9]*$)(?!^[a-zA-Z]*$)^([a-zA-Z0-9]{8,32})$/).test(req.body.pwd))
+          return res.status(400).json({ message: "Erreur d'adresse mail ou de mot de passe." })
         }
         bcrypt.hash(pwd, 10)
           .then(hash => {
-            db.user.create({ firstName: firstName, lastName: lastName, password: hash, email: email })
-              .then(() => { res.status(201).json({ message: "Utilisateur créé" }) })
+            db.user.create({ firstname: firstName, lastname: lastName, password: hash, email: email })
+              .then(() => { return res.status(201).json({ message: "Utilisateur créé" }) })
           })
           .catch(() => {
-            res.status(500).json({ message: "Problème avec le mot de passe" })
+            return res.status(500).json({ message: "Problème avec le mot de passe" })
           })
 
       }
@@ -71,9 +73,7 @@ exports.login = (req, res, next) => {
             /* On coupe le token en 2 afin de stocker une partie en local storage et l'autre en cookie*/
             res.status(200).json({
               user: {
-                idUser: user.idUser, firstName: user.firstName, lastName: user.lastName,
-                email: user.email, telephone: user.telephone, url_avatar: user.url_avatar,
-                description: user.userDescription,
+                idUser: user.idUser,
                 access: user.droits
               },
               token: jwt.sign(
@@ -84,7 +84,7 @@ exports.login = (req, res, next) => {
             });
           })
       }
-    })
+    }).catch(error => res.status(404).json(error))
 };
 
 exports.userList = (req, res, next) => {
@@ -109,64 +109,73 @@ exports.profil = (req, res, next) => {
 
 exports.parametre = (req, res, next) => {
 
-
   const token = req.headers.authorization.split(' ')[1];
   const decodedToken = jwt.verify(token, 'AuheoO11nNej47Gr,eiUHog@ru::ohga5');
   const userId = decodedToken.userId;
 
-  if (userId == req.params.id.split("=")[1]) {
-    if (req.body.pwd1 != "" && req.body.pwd1 === req.body.pwd2) {
-      if (new RegExp(/(?!^[0-9]*$)(?!^[a-zA-Z]*$)^([a-zA-Z0-9]{8,32})$/).test(req.body.pwd1)) {
-        bcrypt.hash(req.body.pwd1, 10)
-          .then(hash => {
-            db.user.update({ password: hash, }, { where: { idUser: userId } })
-              .catch(error => res.status(500).json({ error }))
-          })
+  db.user.findOne({ where: { idUser: userId } })
+    .then(value => {
+      if (value === null) {
+        return res.status(404).json({ message: "Utilisateur inéxistant." })
       }
-    }
+      else {
+        if (userId == req.params.id.split("=")[1]) {
+          if (req.body.pwd1 != "" && req.body.pwd1 === req.body.pwd2) {
+            if (new RegExp(/(?!^[0-9]*$)(?!^[a-zA-Z]*$)^([a-zA-Z0-9]{8,32})$/).test(req.body.pwd1)) {
+              bcrypt.hash(req.body.pwd1, 10)
+                .then(hash => {
+                  db.user.update({ password: hash, }, { where: { idUser: userId } })
+                    .catch(error => res.status(500).json({ error }))
+                })
+            }
+          }
+          let telephone = "0100000000";
+          if (new RegExp(/^0[\d]{9}/).test(req.body.telephone) && req.body.telephone.length === 10) {
+            telephone = req.body.telephone
+          }
+          if (req.file === undefined) {
+            db.user.update({
+              userDescription: req.body.userDescription,
+              telephone: telephone,
+            }, { where: { idUser: userId } }).then(() => res.status(201).json({ message: "Utilisateur modifié" }))
+              .catch(error => res.status(500).json({ error }))
 
-    if (req.file === undefined) {
-      db.user.update({
-        firstName: req.body.firstname,
-        lastName: req.body.lastname,
-        userDescription: req.body.userDescription,
-        telephone: req.body.telephone,
-      }, { where: { idUser: userId } }).then(() => res.status(201).json({ message: "Utilisateur modifié" }))
-        .catch(error => res.status(500).json({ error }))
+          }
+          else {
+            db.user.findOne({
+              where: { idUser: userId }
+            }).then(resultat => {
+              let url_avatar = ""
+              if (resultat.url_avatar != null && resultat.url_avatar != "") {
+                url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
+              }
 
-    }
-    else {
-      db.user.findOne({
-        where: { idUser: userId }
-      }).then(resultat => {
-        let url_avatar=""
-        if(resultat.url_avatar != null && resultat.url_avatar!="")
-        {
-        url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
+              if (fs.existsSync(url_avatar)) {
+                fs.unlink(url_avatar, (err) => {
+                  if (err) throw err;
+                })
+              }
+              let imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+              db.user.update({
+                userDescription: req.body.userDescription,
+                url_avatar: imageUrl,
+                telephone: telephone,
+              }, { where: { idUser: userId } })
+                .then(() => res.status(201).json({ message: "Utilisateur modifié" }))
+                .catch(error => res.status(500).json({ error }))
+
+            })
+          }
+        }
+        else {
+          return res.status(401).json({ message: "Vous n'êtes pas propriétaire du compte" })
         }
 
-        if (fs.existsSync(url_avatar)) {
-          fs.unlink(url_avatar, (err) => {
-            if (err) throw err;
-          })
-        }
-        let imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-        db.user.update({
-          userName: req.body.firstname,
-          lastName: req.body.lastname,
-          userDescription: req.body.userDescription,
-          url_avatar: imageUrl,
-          telephone: req.body.telephone,
-        }, { where: { idUser: userId } })
-          .then(() => res.status(201).json({ message: "Utilisateur modifié" }))
-          .catch(error => res.status(500).json({ error }))
+      }
 
-      })
-    }
-  }
-  else {
-    return res.status(401).json({ message: "Vous n'êtes pas propriétaire du compte" })
-  }
+    })
+
+
 
 }
 
@@ -178,22 +187,29 @@ exports.delete = (req, res, next) => {
 
   if (userId == req.params.id.split("=")[1]) {
     db.user.findOne({
-      attributes: ["url_avatar"],
+      attributes: ["idUser", "url_avatar"],
       where: { idUser: userId }
     })
       .then(resultat => {
-        if (!resultat.url_avatar == "") {
-          let url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
-          if (fs.existsSync(url_avatar)) {
-            fs.unlink(url_avatar, (err) => {
-              if (err) throw err;
-            })
+        let url_avatar = ""
+        if (resultat != null) {
+          if (!resultat.url_avatar == "") {
+            url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
+            if (fs.existsSync(url_avatar)) {
+              fs.unlink(url_avatar, (err) => {
+                if (err) throw err;
+              })
+            }
           }
+          db.user.destroy({ where: { idUser: userId } })
+            .then(() => { return res.status(200).json({ message: "Compte effacé" }) })
+            .catch(error => { return res.status(500).json({ error }) })
+        }
+        else {
+          return res.status(404).json({ message: "Vous voulez supprimer un compte non existant" })
         }
       })
-    db.user.destroy({ where: { idUser: userId } })
-      .then(() => { return res.status(200).json({ message: "Compte effacé" }) })
-      .catch(error => { return res.status(500).json({ error }) })
+
   }
   else {
     return res.status(401).json({ message: "Vous n'êtes pas propriétaire du compte" })
@@ -214,12 +230,14 @@ exports.deleteAvatar = (req, res, next) => {
       where: { idUser: userId }
     })
       .then(resultat => {
-        if (!resultat.url_avatar == "") {
-          let url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
-          if (fs.existsSync(url_avatar)) {
-            fs.unlink(url_avatar, (err) => {
-              if (err) throw err;
-            })
+        if (resultat != null) {
+          if (resultat.url_avatar != "") {
+            let url_avatar = "./images" + resultat.url_avatar.split("/images")[1]
+            if (fs.existsSync(url_avatar)) {
+              fs.unlink(url_avatar, (err) => {
+                if (err) throw err;
+              })
+            }
           }
         }
       })
